# docs/swagger.yaml
basePath: /api/v1
definitions:
  helpers.ErrResponse:
    properties:
      message:
        type: string
    type: object
...
  models.BillOwner:
    properties:
      bank_account:
        type: string
      contact:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Split Bill Swagger Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Split Bill API
  version: "1.0"
paths:
...
  /billowners:
    get:
      description: Get all Bill Owner
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BillOwner'
            type: array
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Get all Bill Owner
      tags:
      - billowners
    post:
      consumes:
      - application/json
      description: Create new bill owner
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.BillOwner'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BillOwner'
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Create new bill owner
      tags:
      - billowners
  /billowners/{id}:
    delete:
      description: Delete bill owner by ID
      parameters:
      - description: data
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BillOwner'
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Delete bill owner by ID
      tags:
      - billowners
    get:
      description: Get Bill Owner by ID
      parameters:
      - description: data
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BillOwner'
            type: array
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Get a Bill Owner by ID
      tags:
      - billowners
    put:
      consumes:
      - application/json
      description: Edit Bill Owner
      parameters:
      - description: data
        in: path
        name: id
        required: true
        type: integer
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.BillOwner'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BillOwner'
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
        "500":
          description: 'Internal Server Error: Server failed to process the request'
          schema:
            $ref: '#/definitions/helpers.ErrResponse'
      summary: Edit Bill Owner
      tags:
      - billowners
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"